+++ Source/Coco.Server/Coco.Framework/SessionManager/Contracts/ISessionRoleManager.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace Coco.Framework.SessionManager.Contracts
 {
-    public interface ISessionRoleManager<TRole> where TRole : IdentityRole<int>
+    public interface ISessionRoleManager<TRole> where TRole : IdentityRole<long>
     {
         //Task<ApplicationRole> FindByNameAsync(string roleName);
         //Task<IList<Claim>> GetClaimsAsync(ApplicationRole role);
